Module fe_compiler::abi::elements

Elements used to define contract ABIs.

Structs

Contract
All public interfaces of a Fe contract.

Event
An event interface.

EventField
A single event field.

FuncInput
A single function input.

FuncOutput
A single function output.

Function
A function interface.

ModuleABIs
Wrapper around a map of contract names to their ABIs.

Enums

FuncType
The type of a public function.

StateMutability
The mutability of a public function.

VarType
The type of an event field or function input or output.