Module fe_compiler::types

Structs

CompiledContract
The artifacts of a compiled contract.

CompiledModule
The artifacts of a compiled module.

Type Definitions

Bytecode
The bytecode of a contract as string object.

ContractName
The name of a Fe contract.

FeModuleAst
The AST of a Fe module.

FeSrc
The source of a Fe module as a static string.

JsonAbi
The ABI of a contract as a string.

NamedAbis
A mapping of contract names and their ABIs.

NamedBytecodeContracts
A mapping of contract names and their bytecode.

NamedContracts
A mapping of contract names and their artifacts.

NamedYulContracts
A mapping of contract names and their Yul IR.

YulIr
The intermediate representation of a contract as a string object.