Function fe_compiler::yul::compile

pub fn compile(context: &Context, module: &FeModuleAst) -> NamedYulContracts

Compiles Fe source code to Yul.

Panics

Any failure to compile an AST to Yul is considered a bug, and thus panics. Invalid ASTs should be caught by an analysis step prior to Yul generation.