Module fe_compiler::yul::runtime::functions::data

Functions

all
Return all data runtime functions

alloc
Allocate a given number of bytes.

alloc_mstoren
Stores a value in a newly allocated memory segment.

avail
Returns the highest available pointer.

bytes_mcopys
Copy memory to a given segment of storage.

bytes_scopym
Copy storage to a newly allocated segment of memory.

bytes_scopys
Copies a segment of storage to another segment of storage.

bytes_sloadn
Read a value of n bytes at the given byte address.

bytes_sstoren
Stores a value in storage at the given address, only modifying a segment of the given size.

ccopym
Copy calldata to a newly allocated segment of memory.

ceil32
Rounds a 256 bit value up to the nearest multiple of 32.

cloadn
Read a value of n bytes from calldata at the given address.

free
Set the highest available pointer.

load_data_string
Load a static string from data into a newly allocated segment of memory.

map_value_ptr
Derives the byte address of a value corresponding to a map key.

mcopym
Copies a segment of memory to another segment of memory.

mcopys
Copy memory to a given segment of storage.

mloadn
Read a value of n bytes from memory at the given address.

mstoren
Stores a value in memory, only modifying the given size (0 < size <= 32).

revert_with_reason_string
Revert with encoded reason string

scopym
Copy storage to a newly allocated segment of memory.

scopys
Copies a segment of storage to another segment of storage.

set_zero
Set the given segment of the value (defined in bits) to zero.

sloadn
Read a value of n bytes at the given word address and bytes offset.

sstoren
Stores a value in storage, only modifying the given size (0 < size <= 32).

ternary
Evaluates the ternary expression and returns the result.